model/serve/requirements.txt
Flask==2.2.3
numpy==1.24.2
Pillow==9.5.0
torch==2.0.0
torchvision==0.15.1
